fix planetary tests
